Made a service for confirm modals

+++ DatingApp/ClientApp/src/app/messages/messages.component.ts
@@ -1,7 +1,8 @@
 import { Component, OnInit } from '@angular/core';
+import {MessageService} from "../_services/message.service";
 import {Message} from "../_models/message";
 import {Pagination} from "../_models/pagination";
-import {MessageService} from "../_services/message.service";
+import {ConfirmService} from "../_services/confirm.service";
 
 @Component({
   selector: 'app-messages',
@@ -16,7 +17,7 @@ export class MessagesComponent implements OnInit {
   pageSize = 5;
   loading = false;
 
-  constructor(private messageService: MessageService) { }
+  constructor(private messageService: MessageService, private confirmService: ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -33,10 +34,14 @@ export class MessagesComponent implements OnInit {
   }
 
   deleteMessage(id: string) {
-    this.messageService.deleteMessage(id)
-      .subscribe(() => {
-        this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
-      })
+    this.confirmService.confirm('Confirm delete message', 'This cannot be undone').subscribe(result => {
+      if (result) {
+        this.messageService.deleteMessage(id)
+          .subscribe(() => {
+            this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
+          })
+      }
+    })
   }
 
   pageChanged(event: any) {
